+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.31.0] - 2024-11-22
 ### Added
 - Added Terms of Service link
+- Hint extension for tile widget account
 
 ## [0.30.0] - 2024-11-22
 ### Added