+++ CHANGELOG.md
@@ -8,11 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Changed
 - Readme section: modified 'Run' button into a link
+- Hint extension for tile widget account
 
 ## [0.31.0] - 2024-11-22
 ### Added
 - Added Terms of Service link
-- Hint extension for tile widget account
 
 ## [0.30.0] - 2024-11-22
 ### Added