+++ CHANGELOG.md
@@ -5,11 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.33.0] - 2024-12-03
 ### Added
 - Warning when deleting datasets which are out of sync with their push remotes
 ### Fixed
 - Visibility for Update button and Flows tab from configuration
+- Error in console during initialization
+- Readme section: redirection after clicking the `Run` button
 
 ## [0.32.0] - 2024-11-28
 ### Added